pyproject.toml
[project]
name = "mine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-samples",
    "fastapi[all]>=0.115.12",
    "jsonrpcclient>=4.0.3",
    "python-dotenv>=1.1.0",
]

[tool.uv.sources]
a2a-samples = { git = "https://github.com/google/A2A", subdirectory = "samples/python" }

